# Conflicts:
#	src/main/java/com/encore/playground/domain/feed/controller/FeedAPIController.java
#	src/main/java/com/encore/playground/domain/notice/controller/NoticeController.java
#	src/main/java/com/encore/playground/domain/qna/controller/AnswerController.java
#	src/main/java/com/encore/playground/domain/qna/controller/QuestionController.java

+++ src/main/java/com/encore/playground/domain/feed/controller/FeedAPIController.java
@@ -39,8 +39,11 @@ public class FeedAPIController {
     @Operation(summary = "피드 메인페이지", description = "메인 페이지 접속 시 피드 목록을 반환한다.")
     @GetMapping(value = "/list")
     public Slice<FeedListDto> feedMain(HttpServletRequest request, @PageableDefault(size=10) Pageable pageable) {
-        MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
-        return feedService.feedPage(memberIdDto, pageable);
+        if (request.getAttribute("AccessTokenValidation").equals("true")) {
+            MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
+            return feedService.feedPage(memberIdDto, pageable);
+        } else
+            return null;
     }
 
     /**
@@ -55,30 +58,33 @@ public class FeedAPIController {
     @Parameter(name = "id", description = "피드 글 번호", example = "1", required = true)
     @GetMapping(value = "/view/{id}")
     public ResponseEntity<?> getFeed(@PathVariable Long id, HttpServletRequest request) {
-        MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
-        HashMap<String, Object> feedAndComments = new HashMap<>();
-        FeedListDto feedListDto = feedService.getFeed(id, memberIdDto);
+        if (request.getAttribute("AccessTokenValidation").equals("true")) {
+            MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
+            HashMap<String, Object> feedAndComments = new HashMap<>();
+            FeedListDto feedListDto = feedService.getFeed(id, memberIdDto);
         feedAndComments.put("feed", feedListDto);
         feedAndComments.put("comments", commentService.getCommentsInFeed(
                         CommentReadDto.builder().feedId(id).build(),
                         memberIdDto,
                         Pageable.ofSize(10)
-                )
-        );
-        String responseMessage = "";
-        if (feedService.isFeedWriter(feedListDto.getId(), memberIdDto)) {
-            responseMessage = ResponseMessage.FEED_DETAIL_SUCCESS;
-        } else {
-            responseMessage = ResponseMessage.FEED_DETAIL_FAILED;
-        }
-        return new ResponseEntity<>(
-                DefaultResponse.res(
-                        StatusCode.OK,
-                        responseMessage,
-                        feedAndComments
-                ),
-                HttpStatus.OK
-        );
+                    )
+            );
+            String responseMessage = "";
+            if (feedService.isFeedWriter(feedListDto.getId(), memberIdDto)) {
+                responseMessage = ResponseMessage.FEED_DETAIL_SUCCESS;
+            } else {
+                responseMessage = ResponseMessage.FEED_DETAIL_FAILED;
+            }
+            return new ResponseEntity<>(
+                    DefaultResponse.res(
+                            StatusCode.OK,
+                            responseMessage,
+                            feedAndComments
+                    ),
+                    HttpStatus.OK
+            );
+        } else
+            return null;
     }
 
     @GetMapping(value = "/test")
@@ -109,8 +115,10 @@ public class FeedAPIController {
     )
     @PostMapping(value = "/write")
     public void write(@RequestBody FeedWriteDto feedWriteDto, HttpServletRequest request) {
-        MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
-        feedService.write(feedWriteDto, memberIdDto);
+        if (request.getAttribute("AccessTokenValidation").equals("true")) {
+            MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
+            feedService.write(feedWriteDto, memberIdDto);
+        }
     }
 
     /**
@@ -131,25 +139,28 @@ public class FeedAPIController {
     )
     @PostMapping(value = "/modify")
     public ResponseEntity<?> modify(@RequestBody FeedModifyDto feedModifyDto, HttpServletRequest request) {
-        MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
-        if (feedService.isFeedWriter(feedModifyDto.getId(), memberIdDto)){
-            feedService.modify(feedModifyDto, memberIdDto);
-            return new ResponseEntity<>(
-                    DefaultResponse.res(
-                            StatusCode.OK,
-                            ResponseMessage.FEED_MODIFY_SUCCESS
-                    ),
-                    HttpStatus.OK
-            );
-        } else {
-            return new ResponseEntity<>(
-                    DefaultResponse.res(
-                            StatusCode.UNAUTHORIZED,
-                            ResponseMessage.FEED_MODIFY_FAILED
-                    ),
-                    HttpStatus.UNAUTHORIZED
-            );
-        }
+        if (request.getAttribute("AccessTokenValidation").equals("true")) {
+            MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
+            if (feedService.isFeedWriter(feedModifyDto.getId(), memberIdDto)) {
+                feedService.modify(feedModifyDto, memberIdDto);
+                return new ResponseEntity<>(
+                        DefaultResponse.res(
+                                StatusCode.OK,
+                                ResponseMessage.FEED_MODIFY_SUCCESS
+                        ),
+                        HttpStatus.OK
+                );
+            } else {
+                return new ResponseEntity<>(
+                        DefaultResponse.res(
+                                StatusCode.UNAUTHORIZED,
+                                ResponseMessage.FEED_MODIFY_FAILED
+                        ),
+                        HttpStatus.UNAUTHORIZED
+                );
+            }
+        } else
+            return null;
     }
 
     /**
@@ -169,24 +180,27 @@ public class FeedAPIController {
     )
     @PostMapping(value = "/delete")
     public ResponseEntity<?> delete(@RequestBody FeedDeleteDto feedDeleteDto, HttpServletRequest request) {
-        MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
-        if (feedService.isFeedWriter(feedDeleteDto.getId(), memberIdDto)){
-            feedService.delete(feedDeleteDto, memberIdDto);
-            return new ResponseEntity<>(
-                    DefaultResponse.res(
-                            StatusCode.OK,
-                            ResponseMessage.FEED_DELETE_SUCCESS
-                    ),
-                    HttpStatus.OK
-            );
-        } else {
-            return new ResponseEntity<>(
-                    DefaultResponse.res(
-                            StatusCode.UNAUTHORIZED,
-                            ResponseMessage.FEED_DELETE_FAILED
-                    ),
-                    HttpStatus.UNAUTHORIZED
-            );
-        }
+        if (request.getAttribute("AccessTokenValidation").equals("true")) {
+            MemberGetMemberIdDto memberIdDto = (MemberGetMemberIdDto) request.getAttribute("memberIdDto");
+            if (feedService.isFeedWriter(feedDeleteDto.getId(), memberIdDto)) {
+                feedService.delete(feedDeleteDto, memberIdDto);
+                return new ResponseEntity<>(
+                        DefaultResponse.res(
+                                StatusCode.OK,
+                                ResponseMessage.FEED_DELETE_SUCCESS
+                        ),
+                        HttpStatus.OK
+                );
+            } else {
+                return new ResponseEntity<>(
+                        DefaultResponse.res(
+                                StatusCode.UNAUTHORIZED,
+                                ResponseMessage.FEED_DELETE_FAILED
+                        ),
+                        HttpStatus.UNAUTHORIZED
+                );
+            }
+        } else
+            return null;
     }
 }
